Last Batch.  Headed South tomorrow
These are the last I can do.  Check the getvar_t function as I have
changed it from the original example which extracted 'time' from a
'datetime' string.  I could not get that to work.  It appears that the
data must be a time value only.  If my writeup is incorrect, please fix
it for me.  Thanks.